simulator/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "helpdesk-simulator"
version = "0.1.0"
description = "Config-driven Zammad training simulator backend"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "HelpDesk Simulator Contributors" }]
dependencies = [
  "apscheduler>=3.10.4,<4.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "pyyaml>=6.0.2,<7.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0,<9.0.0",
  "ruff>=0.6.0,<1.0.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
helpdesk_sim = ["web/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008"]
